add flag to use old adam, needed for backwards compatability
Summary: Old models that use FairseqAdam and eventually use Adam as the internal optimizer can no longer be loaded as checkpoints, because now FairseqAdam seems to always load FusedAdamV1. FusedAdamV1 and Adam are not compatible, meaning all old checkpoints are not loadable with the current code. This fix is to just add a flag to specify if we want to use Adam rather than FusedAdamV1. Later on we'll delete this flag once all models are using FusedAdamV1.

Reviewed By: myleott

Differential Revision: D19358962

fbshipit-source-id: a0af5d50588dc108339a77736dcc8ff5db314dd0