Check if the checkpoint is from the latest version before updating the state_dict in TransformerDecoder.upgrade_state_dict_named() (#2222)
Summary:
Pull Request resolved: https://github.com/pytorch/fairseq/pull/2222

When share_input_output_embed is set to True, the existing code always overrides output_projection.weight with embed_tokens.weight

This is unncessary, and caused a very obscure bug in our custom BART model.

Added a check to skip the update to state_dict if f"{name}.output_projection.weight" is already in the checkpoint.

Reviewed By: myleott

Differential Revision: D21915833

fbshipit-source-id: d298e24394be2ee85c8f686ba459b7e4cbd4298a