Better estimation of shard data sizes for sampling ratios computation in multilingual training
Summary: Previously we use current_data_size * num_shards as the total data size of a sharding dataset. It can be a bad estimation for small shards. Here as the training goes one, multilingual_data_manager will cache the shard size so that it can have exact total data sizes after a pass over the data.

Differential Revision: D23683717

fbshipit-source-id: fcd7fd170327091bfd0bed599f1d930c71f62570