TALNet: Consistency loss
Summary:
During training, for each batch, apply two different data augmentations, and require the model output to be similar. The dissimilarity between the output is used as an extra loss term.

To apply two different augmentations to the same batch, we need to create batches containing two copies of the same indices, like [a, b, c, a, b, c].
I make this the responsibility of the **batch sampler**.
TALNet uses a `DataBalancingEpochBatchIterator` to generate batches. This iterator creates a `_batch_sampler` within itself; this diff adds a `dups` argument to generate duplicate indices.
If you're using the generic `EpochBatchIterator`, which accepts a batch sampler directly as input, you will need to modify the code of your own batch sampler to generate duplicate indices.
This is actually quite easy: if `batch_sampler(n)` creates a generator that returns n indices at a time, then `(batch * 2 for batch in batch_sampler(n // 2))` creates a generator that returns two copies of n/2 indices at a time.

The consistency loss is implemented by the `consistency_loss` function.
To use it, call it in your model's `forward` method and put the result in the return dict.
Then expose the loss in the `get_losses` or `get_extra_losses` method of your model, so criterion objects can read them.

Two types of consistency losses are provided:
1. KL divergence;
2. Squared difference of the probabilities.

I've observed significant gains in TALNet's MAP with both types; KL divergence is slightly better.

Reviewed By: nayansinghal

Differential Revision: D27179650

fbshipit-source-id: 76899335b029fab67bbd7941429c9bd1baf52d65