Do FP16/BF16 conversions on the host to transfer less through PCIe
Summary:
If we do the FP16/BF32 conversion on the host, we do it at DRAM speed but transfer 2X smaller buffer to the GPU through PCIe. PCIe bandwidth is an order of magnitude lower so we actually gain about 50% of execution time compared to when performing the quantization on the GPU.

Also, by transfering an already FP16 buffer, we save memory capacity.

Reviewed By: zhengwy888

Differential Revision: D24146486

fbshipit-source-id: b897e7a32835aa1b571b0fae5f3d72a131ad16a1