Add sequence length padding option for wav2vec2 (#2715)
Summary:
- Adds option to pad encoder input to be a multiple of specified value.
- Speeds up pre-training time for wav2vec2 base model by ~15-20%

# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [x] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/main/CONTRIBUTING.md)?
- [ ] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

## What does this PR do?
Adds option to pad the encoder input to be a multiple of specified value which speeds up fP16 wav2vec2 pretraining by 15-25% for different configurations.

## PR review
Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in Github issues there's a high chance it will not be merged.

## Did you have fun?
Make sure you had fun coding �

Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/2715

Reviewed By: arbabu123

Differential Revision: D33004443

Pulled By: alexeib

fbshipit-source-id: 13141c779f647f05535f66b7a9f3fd1241c8fb58